[core] Eliminate FileCache interface
There is only one implementation and we're unlikely to add more.
